python/snaking/pyproject.toml
[project]
name = "snaking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastdb4py>=0.1.4",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.1",
]
